refactor polynomial model class and deseasonality functions, refine comments
