--- src/Microsoft.ML.TimeSeries/Deseasonality/DeseasonalityBase.cs
@@ -1,7 +1,0 @@
-﻿namespace Microsoft.ML.TimeSeries
-{
-    internal abstract class DeseasonalityBase
-    {
-        public abstract void Deseasonality(ref double[] values, int period, ref double[] results);
-    }
-}
