+++ src/Microsoft.ML.TimeSeries/Microsoft.ML.TimeSeries.csproj
@@ -10,8 +10,4 @@
     <ProjectReference Include="..\Microsoft.ML.Data\Microsoft.ML.Data.csproj" />
     <ProjectReference Include="..\Microsoft.ML.Transforms\Microsoft.ML.Transforms.csproj" />
   </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Deseasonality\" />
-  </ItemGroup>
 </Project>
