+++ src/Microsoft.ML.TimeSeries/STL/LoessBasicParameters.cs
@@ -3,7 +3,7 @@
     internal class LoessBasicParameters
     {
         /// <summary>
-        /// the minimum length of a valid time series. a time series with length equals 2 is so trivial. when less than 2, meaningless.
+        /// The minimum length of a valid time series. A time series with length equals 2 is so trivial and meaningless less than 2.
         /// </summary>
         public const int MinTimeSeriesLength = 3;
     }
